Rename all gateway-tests to gateway-legacy-tests
