+++ tests/application-gateway-legacy-tests/test/helmtest/Dockerfile
@@ -1,6 +1,6 @@
 FROM golang:1.13-alpine as builder
 
-ENV SRC_DIR=/go/src/github.com/kyma-project/kyma/tests/application-gateway-tests
+ENV SRC_DIR=/go/src/github.com/kyma-project/kyma/tests/application-gateway-legacy-tests
 
 WORKDIR $SRC_DIR
 COPY . $SRC_DIR
@@ -11,9 +11,9 @@ FROM alpine:3.10
 
 LABEL source=git@github.com:kyma-project/kyma.git
 
-COPY --from=builder /go/src/github.com/kyma-project/kyma/tests/application-gateway-tests/scripts/helm-test-entrypoint.sh .
-COPY --from=builder /go/src/github.com/kyma-project/kyma/tests/application-gateway-tests/proxyhelmtests.test .
-COPY --from=builder /go/src/github.com/kyma-project/kyma/tests/application-gateway-tests/licenses ./licenses
+COPY --from=builder /go/src/github.com/kyma-project/kyma/tests/application-gateway-legacy-tests/scripts/helm-test-entrypoint.sh .
+COPY --from=builder /go/src/github.com/kyma-project/kyma/tests/application-gateway-legacy-tests/proxyhelmtests.test .
+COPY --from=builder /go/src/github.com/kyma-project/kyma/tests/application-gateway-legacy-tests/licenses ./licenses
 
 ARG TEST_EXECUTOR_IMAGE
 ENV TEST_EXECUTOR_IMAGE=$TEST_EXECUTOR_IMAGE
